requirements-dev.txt
asttokens==2.4.1
coverage==7.4.3
decorator==5.1.1
deprecation==2.1.0
executing==2.0.1
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.22.2
jedi==0.19.1
matplotlib-inline==0.1.6
numpy==1.26.4
packaging==24.0
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pytest==8.1.1
pytest-cov==4.1.0
pytest-random-order==1.1.1
PyYAML==6.0.1
redis==5.0.3
ruff==0.3.3
six==1.16.0
stack-data==0.6.3
toml-sort==0.23.1
tomlkit==0.12.4
traitlets==5.14.2
wcwidth==0.2.13
